* Now thats why it doesn't work
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@2040 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ player.cpp
@@ -3334,8 +3334,8 @@ uint64_t Player::getGainedExperience(Creature* attacker, bool useMultiplier/* = 
 	if(!attackerPlayer || attackerPlayer == this)
 		return 0;
 
-	uint32_t attackerLevel = attackerPlayer->getLevel(), min = (int32_t)(g_config.getDouble(
-		ConfigManager::EFP_MIN_THRESHOLD)), max = (int32_t)(g_config.getDouble(ConfigManager::EFP_MAX_THRESHOLD));
+	double attackerLevel = (double)attackerPlayer->getLevel(), min = g_config.getDouble(
+		ConfigManager::EFP_MIN_THRESHOLD), max = g_config.getDouble(ConfigManager::EFP_MAX_THRESHOLD);
 	if((min > 0 && level < (uint32_t)std::floor(attackerLevel * min)) || (max > 0 &&
 		level > (uint32_t)std::floor(attackerLevel * max)))
 		return 0;
@@ -3349,13 +3349,13 @@ uint64_t Player::getGainedExperience(Creature* attacker, bool useMultiplier/* = 
 		result = (1 - (a / b)) * 0.05 * c
 	*/
 	uint32_t a = (uint32_t)std::floor(attackerLevel * 0.9), b = level;
-	uint64_t c = getExperience(), result = std::max((uint64_t)0, (uint64_t)std::floor(
-		getDamageRatio(attacker) * std::max((double)0, ((double)(1 - (((double)a / b))))) * 0.05 * c));
+	uint64_t c = getExperience(), result = std::max((uint64_t)0, (uint64_t)std::floor(getDamageRatio(
+		attacker) * std::max((double)0, ((double)(1 - (((double)a / b))))) * 0.05 * c));
 	if(useMultiplier)
 		result = uint64_t((double)result * attackerPlayer->rates[SKILL__LEVEL]);
 
-	return std::min((uint64_t)getLostExperience(), uint64_t(result * g_game.getExperienceStage(
-		attackerLevel, attackerPlayer->getVocation()->getExperienceMultiplier())));
+	return std::min((uint64_t)getLostExperience(), (uint64_t)(result * g_game.getExperienceStage(
+		(uint32_t)attackerLevel, attackerPlayer->getVocation()->getExperienceMultiplier())));
 }
 
 void Player::onFollowCreature(const Creature* creature)
